CHanged pause controller, pause pannel
