+++ Assets/Scenes/SampleScene.unity
@@ -870,6 +870,7 @@ GameObject:
   - component: {fileID: 1689673601}
   - component: {fileID: 1689673604}
   - component: {fileID: 1689673603}
+  - component: {fileID: 1689673605}
   m_Layer: 5
   m_Name: PausePannel
   m_TagString: Untagged
@@ -934,6 +935,19 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1689673600}
   m_CullTransparentMesh: 1
+--- !u!114 &1689673605
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1689673600}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 777321208188f5a4e91de16aa2b6bac2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _pauseController: {fileID: 868551406}
 --- !u!1 &1778819143
 GameObject:
   m_ObjectHideFlags: 0
