--- Assets/Scripts/PlayerMechanics/PlayerPause.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-public class PlayerPause : MonoBehaviour
-{
-    [SerializeField] private GameObject pausePannel;
-    public void Pause()
-    {
-        pausePannel.SetActive(!pausePannel.activeSelf);
-    }
-}
